Rename queries to api
