+++ src/server/routes/Supplier.js
@@ -1,10 +1,10 @@
-const SupplierApi = require('../queries/Supplier');
-const SupplierVisibilityApi = require('../queries/SupplierVisibility');
-const SupplierBankAccountApi = require('../queries/SupplierBankAccount');
-const SupplierAddressApi = require('../queries/SupplierAddress');
-const SupplierContactApi = require('../queries/SupplierContact');
-const SupplierCapabilityApi = require('../queries/Capability');
-const Supplier2UserApi = require('../queries/Supplier2User');
+const SupplierApi = require('../api/Supplier');
+const SupplierVisibilityApi = require('../api/SupplierVisibility');
+const SupplierBankAccountApi = require('../api/SupplierBankAccount');
+const SupplierAddressApi = require('../api/SupplierAddress');
+const SupplierContactApi = require('../api/SupplierContact');
+const SupplierCapabilityApi = require('../api/Capability');
+const Supplier2UserApi = require('../api/Supplier2User');
 const userService = require('../services/user');
 const businessLinkService = require('../services/businessLink');
 const Promise = require('bluebird');
